# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "GlowMart API",
        "description": "REST API E-Commerce Makeup GlowMart — JWT Auth, Role-based Access, API Gateway, Third-party Makeup API Integration",
        "contact": {
            "email": "admin@glowmart.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Dev"
        }
    ],
    "paths": {
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login dan dapatkan JWT token",
                "operationId": "b2d79c995139fbe1bbd02d802c8ff735",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "admin@glowmart.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login berhasil"
                    },
                    "401": {
                        "description": "Kredensial salah"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                }
            }
        },
        "/api/auth/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Registrasi akun baru",
                "operationId": "06bb6d28c6f72090906e933363873669",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Siti Aminah"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "siti@email.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password123"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Registrasi berhasil"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                }
            }
        },
        "/api/auth/profile": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Ambil profil user yang sedang login",
                "operationId": "2ed31709dc1b2613897f5beedec4b521",
                "responses": {
                    "200": {
                        "description": "Berhasil"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Daftar semua produk aktif (public)",
                "operationId": "d53e9ba5638f83b411c003721d0dd0ac",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "category_id",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "brand",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "min_price",
                        "in": "query",
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "max_price",
                        "in": "query",
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil"
                    }
                }
            }
        },
        "/api/cart/add": {
            "post": {
                "tags": [
                    "Cart"
                ],
                "summary": "Tambah produk ke keranjang",
                "operationId": "80439e6e41d8711ac96b537e8a32df38",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "product_id",
                                    "quantity"
                                ],
                                "properties": {
                                    "product_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "quantity": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Berhasil ditambahkan"
                    },
                    "400": {
                        "description": "Stok tidak cukup"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/orders": {
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Buat pesanan baru dari keranjang",
                "operationId": "4f5424496a658534f31f9a46210d9455",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "shipping_address",
                                    "shipping_city",
                                    "payment_method"
                                ],
                                "properties": {
                                    "shipping_address": {
                                        "type": "string",
                                        "example": "Jl. Contoh No. 1"
                                    },
                                    "shipping_city": {
                                        "type": "string",
                                        "example": "Malang"
                                    },
                                    "payment_method": {
                                        "type": "string",
                                        "enum": [
                                            "transfer",
                                            "cod",
                                            "ewallet"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Pesanan berhasil dibuat"
                    },
                    "400": {
                        "description": "Keranjang kosong"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/external/makeup/search": {
            "get": {
                "tags": [
                    "External"
                ],
                "summary": "Cari produk dari Makeup API",
                "operationId": "f2235ff2fb7225344fedacce8f8bc89e",
                "parameters": [
                    {
                        "name": "brand",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "product_type",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil dari API eksternal"
                    }
                }
            }
        },
        "/api/admin/dashboard": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Statistik dashboard admin",
                "operationId": "33fbf6cd7ab49a5d9076f55e4755dcbb",
                "responses": {
                    "200": {
                        "description": "Berhasil"
                    },
                    "403": {
                        "description": "Akses ditolak"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/shipping/calculate": {
            "post": {
                "tags": [
                    "Shipping"
                ],
                "summary": "Kalkulasi ongkos kirim",
                "operationId": "6ca5be61b5e9cb73eefe6f3b63642563",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "origin",
                                    "destination",
                                    "weight",
                                    "courier"
                                ],
                                "properties": {
                                    "origin": {
                                        "type": "string",
                                        "example": "455"
                                    },
                                    "destination": {
                                        "type": "string",
                                        "example": "455"
                                    },
                                    "weight": {
                                        "type": "integer",
                                        "example": 500
                                    },
                                    "courier": {
                                        "type": "string",
                                        "enum": [
                                            "jne",
                                            "tiki",
                                            "pos"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Berhasil"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Auth",
            "description": "Authentication endpoints"
        },
        {
            "name": "Products",
            "description": "Produk endpoints"
        },
        {
            "name": "Cart",
            "description": "Keranjang belanja"
        },
        {
            "name": "Orders",
            "description": "Pesanan"
        },
        {
            "name": "Admin",
            "description": "Admin only endpoints"
        },
        {
            "name": "External",
            "description": "Third-party API integrations"
        },
        {
            "name": "Shipping",
            "description": "Kalkulasi ongkos kirim"
        }
    ]
}